@@ -1,6 +1,6 @@
 class Admin::ArtistsController < Admin::AdminController
-	def index
-		@artists = Artist.all
+  def index
+    @artists = Artist.all
   end
 
   def new
@@ -8,28 +8,34 @@ class Admin::ArtistsController < Admin::AdminController
   end
 
   def create
-    @artist = Artist.new params[:artist]
+    @artist = Artist.new artist_params
     if @artist.save
-      redirect_to admin_artists_path, :notice => "Successfully Created Artist."
+      redirect_to admin_artists_path, notice: 'Successfully Created Artist.'
     else
       render :new
     end
   end
 
-	def show
-		@artist = Artist.find(params[:id])
+  def show
+    @artist = Artist.find(params[:id])
   end
 
-	def update
-		@artist = Artist.find(params[:id])
-		@artist.name = params[:artist][:name]
-		@artist.save
-		redirect_to :back
-	end
+  def update
+    @artist = Artist.find(params[:id])
+    @artist.name = params[:artist][:name]
+    @artist.save
+    redirect_to :back
+  end
 
   def destroy
     @artist = Artist.find(params[:id])
     @artist.destroy
     redirect_to admin_artists_path
   end
-end	
+
+  protected
+
+  def artist_params
+    params.require(:artist).permit(:name)
+  end
+end
