@@ -1,7 +1,7 @@
 class Admin::CommentsController < Admin::AdminController
   before_filter :load_art
-  before_filter :load_comment, :only => [:unapprove, :approve]
-  
+  before_filter :load_comment, only: [:unapprove, :approve]
+
   def unapprove
     @comment.approved = false
     @comment.save!
@@ -14,24 +14,19 @@ class Admin::CommentsController < Admin::AdminController
     head 200
   end
 
-  def delete   
-   @comment = Comment.find(params[:id])
-   @comment.destroy
-   head 200
+  def delete
+    @comment = Comment.find(params[:id])
+    @comment.destroy
+    head 200
   end
-  
+
   protected
-  
+
   def load_art
-    unless params[:art_id] and (@art = Art.where(:slug => params[:art_id]).first)
-      head :not_found and return false
-    end
+    @art = Art.find(params[:art_id])
   end
-  
+
   def load_comment
-    unless params[:id] and (@comment = @art.comments.where(:_id => BSON::ObjectId(params[:id])).first)
-      head :not_found and return false
-    end
+    @comment = @art.comments.find(params[:id])
   end
-  
 end
