@@ -6,7 +6,8 @@ class Admin::CommissionersController < Admin::AdminController
   def update
     @commissioner.attributes = params[:commissioner]
     if @commissioner.save
-      redirect_to admin_commissioner_path(@commissioner), :notice => "Successfully updated commissioner."
+      redirect_to admin_commissioner_path(@commissioner),
+                  notice: 'Successfully updated commissioner.'
     else
       render :show
     end
@@ -18,15 +19,12 @@ class Admin::CommissionersController < Admin::AdminController
   end
 
   def new
-
   end
 
   protected
 
   def load_commissioner
-    unless params[:id] and (@commissioner = Commissioner.find(params[:id]))
-      head :not_found and return false
-    end
+    @commissioner = Commissioner.find(params[:id])
     @commissioners = Commissioner.all
   end
 end
