@@ -1,6 +1,6 @@
 class Admin::CountriesController < Admin::AdminController
-	def index
-		@countries = Country.all
+  def index
+    @countries = Country.all
   end
 
   def new
@@ -10,26 +10,26 @@ class Admin::CountriesController < Admin::AdminController
   def create
     @country = Country.new params[:country]
     if @country.save
-      redirect_to admin_countries_path, :notice => "Successfully Created Country."
+      redirect_to admin_countries_path, notice: 'Successfully Created Country.'
     else
       render :new
     end
   end
 
-	def show
-		@country = Country.find(params[:id])
+  def show
+    @country = Country.find(params[:id])
   end
 
-	def update
-		@country = Country.find(params[:id])
-		@country.country_count = params[:country][:country_count]
-		@country.save
-		redirect_to :back
-	end
+  def update
+    @country = Country.find(params[:id])
+    @country.country_count = params[:country][:country_count]
+    @country.save
+    redirect_to :back
+  end
 
   def destroy
     @country = Country.find(params[:id])
     @country.destroy
     redirect_to admin_countries_path
   end
-end	
+end
