@@ -3,9 +3,9 @@ class Photo
   include Mongoid::Timestamps
   include Mongoid::Paperclip
 
-  belongs_to :art
+  belongs_to :art, optional: true
 
-  attr_protected :_id
+  # attr_protected :_id
 
   field :attribution_text, :type => String
   field :attribution_url, :type => String
@@ -26,7 +26,8 @@ class Photo
 
   validates_attachment_presence :image, :message => "Please include a photo."
   validates_attachment_size :image, :in => 0..6.megabytes, :message => "Photo must be less than 6 megabytes."
-  validates_attachment_content_type :image, :content_type => ["application/octet-stream", "image/jpeg", "image/gif", "image/jpg", "image/png"], :message => "Please include a JPG, GIF or PNG image for a photo."
+  # validates_attachment_content_type :image, :content_type => ["application/octet-stream", "image/jpeg", "image/gif", "image/jpg", "image/png"], :message => "Please include a JPG, GIF or PNG image for a photo."
+  validates_attachment_content_type :image, content_type: /\Aimage\/.*\z/
 
   def attribution_text_for_display
     return attribution_text unless attribution_text.blank?
